Remove duplicat config eobject
