+++ app/uk/gov/hmrc/bforms/config/frontendAppConfig.scala
@@ -20,10 +20,11 @@ import play.api.Play.{configuration, current}
 import uk.gov.hmrc.play.config.ServicesConfig
 
 trait AppConfig {
-  val analyticsToken: String
-  val analyticsHost: String
-  val reportAProblemPartialUrl: String
-  val reportAProblemNonJSUrl: String
+  def assetsPrefix: String
+  def analyticsToken: String
+  def analyticsHost: String
+  def reportAProblemPartialUrl: String
+  def reportAProblemNonJSUrl: String
   def governmentGatewaySignInUrl: String
   def bFormsFrontendBaseUrl: String
 }
@@ -35,6 +36,7 @@ object FrontendAppConfig extends AppConfig with ServicesConfig {
   private val contactHost = configuration.getString(s"contact-frontend.host").getOrElse("")
   private val contactFormServiceIdentifier = "MyService"
 
+  override lazy val assetsPrefix = loadConfig(s"assets.url") + loadConfig(s"assets.version")
   override lazy val analyticsToken = loadConfig(s"google-analytics.token")
   override lazy val analyticsHost = loadConfig(s"google-analytics.host")
   override lazy val reportAProblemPartialUrl = s"$contactHost/contact/problem_reports_ajax?service=$contactFormServiceIdentifier"
