--- app/uk/gov/hmrc/bforms/controllers/frontendAppConfig.scala
@@ -1,51 +1,0 @@
-/*
- * Copyright 2016 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.bforms.controllers
-
-import play.api.Play.{ configuration, current }
-import uk.gov.hmrc.play.config.ServicesConfig
-
-trait AppConfig {
-  def assetsPrefix: String
-  def analyticsToken: String
-  def analyticsHost: String
-  def reportAProblemPartialUrl: String
-  def reportAProblemNonJSUrl: String
-  def governmentGatewaySignInUrl: String
-  def bFormsFrontendBaseUrl: String
-}
-
-object FrontendAppConfig extends AppConfig with ServicesConfig {
-
-  private def loadConfig(key: String) = configuration.getString(key).getOrElse(throw new Exception(s"Missing configuration key: $key"))
-
-  private val contactHost = configuration.getString(s"contact-frontend.host").getOrElse("")
-  private val contactFormServiceIdentifier = "MyService"
-
-  override lazy val assetsPrefix = loadConfig(s"assets.url") + loadConfig(s"assets.version")
-  override lazy val analyticsToken = loadConfig(s"google-analytics.token")
-  override lazy val analyticsHost = loadConfig(s"google-analytics.host")
-  override lazy val reportAProblemPartialUrl = s"$contactHost/contact/problem_reports_ajax?service=$contactFormServiceIdentifier"
-  override lazy val reportAProblemNonJSUrl = s"$contactHost/contact/problem_reports_nonjs?service=$contactFormServiceIdentifier"
-
-  override lazy val governmentGatewaySignInUrl = configuration.getString("government-gateway-sign-in-url").getOrElse("")
-  //  override lazy val governmentGatewaySignInUrl = "http://localhost:9025/gg/sign-in"
-
-  // this will be empty in non-local environments
-  override lazy val bFormsFrontendBaseUrl = configuration.getString("bform-frontend-base-url").getOrElse("")
-
-}
